# shared/types/src/api-client/models/CreateDoseRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateDoseRequest = {
    medicationId: string;
    status?: CreateDoseRequest.status;
};

export namespace CreateDoseRequest {

    export enum status {
        TAKEN = 'taken',
    }


}