# shared/types/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"DefaultSelection_Prisma._36_MedicationPayload_": {
				"properties": {
					"careRecipientId": {
						"type": "string"
					},
					"instructions": {
						"type": "string"
					},
					"dosage": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"isActive": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"careRecipientId",
					"instructions",
					"dosage",
					"updatedAt",
					"createdAt",
					"isActive",
					"id",
					"name"
				],
				"type": "object"
			},
			"Medication": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_MedicationPayload_",
				"description": "Model Medication"
			},
			"DefaultSelection_Prisma._36_CareRecipientPayload_": {
				"properties": {
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"isActive": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"dateOfBirth",
					"lastName",
					"firstName",
					"updatedAt",
					"createdAt",
					"isActive",
					"id"
				],
				"type": "object"
			},
			"CareRecipient": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_CareRecipientPayload_",
				"description": "Model CareRecipient"
			},
...
			"DoseResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"scheduledFor": {
						"type": "string",
						"format": "date-time"
					},
					"takenAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"medicationId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"medication": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Medication"
							},
							{
								"properties": {
									"careRecipient": {
										"$ref": "#/components/schemas/CareRecipient"
									}
								},
								"type": "object"
							}
						]
					}
				},
				"required": [
					"status",
					"scheduledFor",
					"takenAt",
					"updatedAt",
					"createdAt",
					"medicationId",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateDoseStatusRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"scheduled",
							"taken",
							"missed",
							"skipped"
						]
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateDoseRequest": {
				"properties": {
					"medicationId": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"taken"
						],
						"nullable": false
					}
				},
				"required": [
					"medicationId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "medication_management_api",
		"version": "1.0.0",
		"description": "Medication Management API",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Thomas Bowen"
		}
	},
	"paths": {
...
		"/doses": {
			"get": {
				"operationId": "GetDoses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DoseResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all doses with optional filtering",
				"tags": [
					"Doses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "recipientId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDose",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DoseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Medication not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new dose record",
				"tags": [
					"Doses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDoseRequest"
							}
						}
					}
				}
			}
		},
		"/doses/{id}": {
			"get": {
				"operationId": "GetDose",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DoseResponse"
								}
							}
						}
					},
					"404": {
						"description": "Dose not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a dose by ID",
				"tags": [
					"Doses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/doses/{id}/status": {
			"put": {
				"operationId": "UpdateDoseStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DoseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Dose not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Update a dose status",
				"tags": [
					"Doses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDoseStatusRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}